Attempt saving documents which don't publish themselves as dirty.
* src/Ankh.Services/Commands/CommandMapper.cs
  Fixed compile issue
  
* src/Ankh.Scc/OpenDocumentTracker.Api.cs
  Attempt saving if not dirty
  
* src/Ankh.Scc/ProjectMap/SccDocumentData.cs
  Don't unset dirty in CheckDirty.. The save will fix wrong issues; but this breaks dirty on projects

+++ trunk/src/Ankh.Services/Commands/CommandMapper.cs
@@ -257,7 +257,7 @@ namespace Ankh.Commands
                     cmdf |= OLECMDF.OLECMDF_INVISIBLE;
 
                 if (updateArgs.DynamicMenuEnd)
-                    return OLEConstants.OLECMDERR_E_NOTSUPPORTED;
+                    return (int)OLEConstants.OLECMDERR_E_NOTSUPPORTED;
             }
 
             if (textQuery != TextQueryType.None)
